+++ src/index.js
@@ -1,3 +1,36 @@
+import path from 'path';
+
+import glob from 'glob';
+import {plugin} from 'postcss';
+
+import transformer from './lib/transformer';
+
+export default plugin('postcss-functions', (opts = {}) => {
+	const functions = opts.functions || {};
+	let globs = opts.glob || [];
+
+	if (!Array.isArray(globs))
+		globs = [globs];
+
+	globs.forEach(pattern => {
+		glob.sync(pattern).forEach(file => {
+			const name = path.basename(file, path.extname(file));
+			functions[name] = require(file);
+		});
+	});
+
+	const transform = transformer(functions);
+
+	return css => {
+		const promises = [];
+		css.walk(node => {
+			promises.push(transform(node));
+		});
+		return Promise.all(promises);
+	};
+});
+
+/*
 var postcss = require('postcss'),
 	assign = require('object-assign'),
 	glob = require('glob'),
@@ -33,3 +66,4 @@ module.exports = postcss.plugin('postcss-functions', function (opts) {
 		return Promise.all(promises);
 	};
 });
+*/